requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
pyserial==3.5
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
python-dateutil==2.9.0.post0
shiboken6==6.7.0
six==1.16.0
